Move the getop function to utils
